+++ test/with_dummyserver/async/test_poolmanager.py
@@ -3,12 +3,17 @@ import json
 import pytest
 from ahip.base import DEFAULT_PORTS
 from ahip import PoolManager, Retry
-from ahip.exceptions import MaxRetryError, UnrewindableBodyError
+from ahip.exceptions import MaxRetryError, NewConnectionError, UnrewindableBodyError
+from hip.util.retry import RequestHistory
 
 from test.with_dummyserver import conftest
 from dummyserver.testcase import HTTPDummyServerTestCase
 
 from test import LONG_TIMEOUT
+
+
+def _test_name(test_name, backend, anyio_backend):
+    return "-".join([test_name, backend, anyio_backend or "None"])
 
 
 class TestPoolManager(HTTPDummyServerTestCase):
@@ -410,7 +415,9 @@ class TestFileUploads(HTTPDummyServerTestCase):
         data = b"A" * content_length
         uploaded_file = io.BytesIO(data)
         headers = {
-            "test-name": "test_retries_put_filehandle",
+            "test-name": _test_name(
+                "test_retries_put_filehandle", backend, anyio_backend
+            ),
             "Content-Length": str(content_length),
         }
 
@@ -465,3 +472,241 @@ class TestFileUploads(HTTPDummyServerTestCase):
                 "An error occurred when rewinding request body for redirect/retry."
                 == str(e.value)
             )
+
+
+class TestRetry(HTTPDummyServerTestCase):
+    @classmethod
+    def setup_class(self):
+        super(TestRetry, self).setup_class()
+        self.base_url = "http://%s:%d" % (self.host, self.port)
+        self.base_url_alt = "http://%s:%d" % (self.host_alt, self.port)
+
+    @conftest.test_all_backends
+    async def test_max_retry(self, backend, anyio_backend):
+        with PoolManager(backend=backend) as http:
+            with pytest.raises(MaxRetryError):
+                await http.request(
+                    "GET",
+                    "%s/redirect" % self.base_url,
+                    fields={"target": "/"},
+                    retries=0,
+                )
+
+    @conftest.test_all_backends
+    async def test_disabled_retry(self, backend, anyio_backend):
+        """ Disabled retries should disable redirect handling. """
+        with PoolManager(backend=backend) as http:
+            r = await http.request(
+                "GET",
+                "%s/redirect" % self.base_url,
+                fields={"target": "/"},
+                retries=False,
+            )
+            assert r.status == 303
+
+            r = await http.request(
+                "GET",
+                "%s/redirect" % self.base_url,
+                fields={"target": "/"},
+                retries=Retry(redirect=False),
+            )
+            assert r.status == 303
+
+            with pytest.raises(NewConnectionError):
+                await http.request(
+                    "GET",
+                    "http://thishostdoesnotexist.invalid/",
+                    timeout=0.001,
+                    retries=False,
+                )
+
+    @conftest.test_all_backends
+    async def test_read_retries(self, backend, anyio_backend):
+        """ Should retry for status codes in the whitelist """
+        retry = Retry(read=1, status_forcelist=[418])
+
+        with PoolManager(backend=backend) as http:
+            resp = await http.request(
+                "GET",
+                "%s/successful_retry" % self.base_url,
+                headers={
+                    "test-name": _test_name("test_read_retries", backend, anyio_backend)
+                },
+                retries=retry,
+            )
+            assert resp.status == 200
+
+    @conftest.test_all_backends
+    async def test_read_total_retries(self, backend, anyio_backend):
+        """ HTTP response w/ status code in the whitelist should be retried """
+        headers = {
+            "test-name": _test_name("test_read_total_retries", backend, anyio_backend)
+        }
+        retry = Retry(total=1, status_forcelist=[418])
+
+        with PoolManager(backend=backend) as http:
+            resp = await http.request(
+                "GET",
+                "%s/successful_retry" % self.base_url,
+                headers=headers,
+                retries=retry,
+            )
+            assert resp.status == 200
+
+    @conftest.test_all_backends
+    async def test_retries_wrong_whitelist(self, backend, anyio_backend):
+        """HTTP response w/ status code not in whitelist shouldn't be retried"""
+        headers = {
+            "test-name": _test_name("test_wrong_whitelist", backend, anyio_backend)
+        }
+        retry = Retry(total=1, status_forcelist=[202])
+
+        with PoolManager(backend=backend) as http:
+            resp = await http.request(
+                "GET",
+                "%s/successful_retry" % self.base_url,
+                retries=retry,
+                headers=headers,
+            )
+            assert resp.status == 418
+
+    @conftest.test_all_backends
+    async def test_default_method_whitelist_retried(self, backend, anyio_backend):
+        """Hip should retry methods in the default method whitelist"""
+        headers = {
+            "test-name": _test_name("test_default_whitelist", backend, anyio_backend)
+        }
+        retry = Retry(total=1, status_forcelist=[418])
+
+        with PoolManager(backend=backend) as http:
+            resp = await http.request(
+                "OPTIONS",
+                "%s/successful_retry" % self.base_url,
+                headers=headers,
+                retries=retry,
+            )
+            assert resp.status == 200
+
+    @conftest.test_all_backends
+    async def test_retries_wrong_method_list(self, backend, anyio_backend):
+        """Method not in our whitelist should not be retried, even if code matches"""
+        headers = {
+            "test-name": _test_name(
+                "test_wrong_method_whitelist", backend, anyio_backend
+            )
+        }
+        retry = Retry(total=1, status_forcelist=[418], method_whitelist=["POST"])
+
+        with PoolManager(backend=backend) as http:
+            resp = await http.request(
+                "GET",
+                "%s/successful_retry" % self.base_url,
+                headers=headers,
+                retries=retry,
+            )
+            assert resp.status == 418
+
+    @conftest.test_all_backends
+    async def test_read_retries_unsuccessful(self, backend, anyio_backend):
+        headers = {
+            "test-name": _test_name(
+                "test_read_retries_unsuccessful", backend, anyio_backend
+            )
+        }
+
+        with PoolManager(backend=backend) as http:
+            resp = await http.request(
+                "GET", "%s/successful_retry" % self.base_url, headers=headers, retries=1
+            )
+            assert resp.status == 418
+
+    @conftest.test_all_backends
+    async def test_retry_reuse_safe(self, backend, anyio_backend):
+        """ It should be possible to reuse a Retry object across requests """
+        headers = {"test-name": _test_name("test_retry_safe", backend, anyio_backend)}
+        retry = Retry(total=1, status_forcelist=[418])
+
+        with PoolManager(backend=backend) as http:
+            resp = await http.request(
+                "GET",
+                "%s/successful_retry" % self.base_url,
+                headers=headers,
+                retries=retry,
+            )
+            assert resp.status == 200
+            resp = await http.request(
+                "GET",
+                "%s/successful_retry" % self.base_url,
+                headers=headers,
+                retries=retry,
+            )
+            assert resp.status == 200
+
+    @conftest.test_all_backends
+    async def test_retry_return_in_response(self, backend, anyio_backend):
+        headers = {
+            "test-name": _test_name(
+                "test_retry_return_in_response", backend, anyio_backend
+            )
+        }
+        retry = Retry(total=2, status_forcelist=[418])
+
+        with PoolManager(backend=backend) as http:
+            resp = await http.request(
+                "GET",
+                "%s/successful_retry" % self.base_url,
+                headers=headers,
+                retries=retry,
+            )
+            assert resp.status == 200
+            assert resp.retries.total == 1
+            assert resp.retries.history == (
+                RequestHistory("GET", "/successful_retry", None, 418, None),
+            )
+
+    @conftest.test_all_backends
+    async def test_retry_redirect_history(self, backend, anyio_backend):
+        with PoolManager(backend=backend) as http:
+            resp = await http.request(
+                "GET", "%s/redirect" % self.base_url, fields={"target": "/"}
+            )
+            assert resp.status == 200
+            assert resp.retries.history == (
+                RequestHistory(
+                    "GET", self.base_url + "/redirect?target=%2F", None, 303, "/"
+                ),
+            )
+
+    @conftest.test_all_backends
+    async def test_multi_redirect_history(self, backend, anyio_backend):
+        with PoolManager(backend=backend) as http:
+            r = await http.request(
+                "GET",
+                "%s/multi_redirect" % self.base_url,
+                fields={"redirect_codes": "303,302,200"},
+                redirect=False,
+            )
+            assert r.status == 303
+            assert r.retries.history == tuple()
+
+            r = await http.request(
+                "GET",
+                "%s/multi_redirect" % self.base_url,
+                retries=10,
+                fields={"redirect_codes": "303,302,301,307,302,200"},
+            )
+            assert r.status == 200
+            assert r.data == b"Done redirecting"
+
+            expected = [
+                (303, "/multi_redirect?redirect_codes=302,301,307,302,200"),
+                (302, "/multi_redirect?redirect_codes=301,307,302,200"),
+                (301, "/multi_redirect?redirect_codes=307,302,200"),
+                (307, "/multi_redirect?redirect_codes=302,200"),
+                (302, "/multi_redirect?redirect_codes=200"),
+            ]
+            actual = [
+                (history.status, history.redirect_location)
+                for history in r.retries.history
+            ]
+            assert actual == expected
